feature("cs:namespace")
This feature, for csharp, lets users indicate exactly what namespace a class
is in, in .Net.  This is useful if you're using multiple modules, each
in their own namespace, and some of them refer to types from other modules.

This causes those other modules (and the original module) to refer
to these types with their full namespace.  Without this, the only fix was
to add a using declaration in those outside modules, which can lead to
ambiguity if you have types of the same name in both namespaces.